final_2

+++ vue/src/views/Search.vue
@@ -113,7 +113,9 @@ export default {
     },
     // TODO 管理者関連のの記述
     deleteGraphData (parentId) {
-      this.$store.dispatch('search/deleteGraphData', parentId)
+      if (confirm('本当に削除してよろしいですか？')) {
+        this.$store.dispatch('search/deleteGraphData', parentId)
+      }
     }
   }
 }
